+++ src/bootstrap.scss
@@ -38,6 +38,9 @@ h1,h2,h3,h4,h5, button, .heading-type {
 
 h1 {
 	margin-top: 1.5rem;
+	&:first-child{
+		margin-top:0;
+	}
 	margin-bottom: 2rem;
 
 	@include media-breakpoint-down(xs) {
