+++ src/ui/compose/compose.scss
@@ -24,6 +24,7 @@
 
 	.bb-compose-song-player {
 		flex-grow: 1;
+		padding: .5em 1em;
 		height: 100%;
 		width: 0;
 	}
