+++ src/ui/overview/layout.scss
@@ -1,4 +1,7 @@
 .bb-overview {
+	.h-100 {
+		height: 100vh !important;
+	}
 
 	.bb-overview-toggle-patternList {
 		position: absolute;
@@ -41,11 +44,9 @@
 			a { 
 				border-bottom: 2px solid transparent; 
 			}
-			&.active, &:hover {
-				a {
-					border-color: black;
-					text-decoration:none;
-				}
+			.router-link-active.with-child-links, .router-link-exact-active, &:hover a {
+				border-color: black;
+				text-decoration:none;
 			}
 		}
 	}
