+++ src/ui/pattern-placeholder/pattern-placeholder.scss
@@ -10,10 +10,6 @@
 		visibility: hidden;
 	}
 
-	.card-body {
-		padding: 0;
-	}
-
 	.pattern-button {
 		height: 100%;
 		box-sizing: border-box;
