+++ src/ui/pattern-placeholder/pattern-placeholder.vue
@@ -9,7 +9,7 @@
 	<ul class="actions icon-list">
 		<li><a href="javascript:" title="Listen" v-b-tooltip.hover @click="playPattern()"><fa :icon="playerRef && playerRef.playing ? 'stop' : 'play-circle'"></fa></a></li>
 		<li><router-link
-			:to="{ name: 'listen pattern', params: { tuneName, patternName } }"
+			:to="{ name: readonly ? 'listen pattern' : 'edit pattern', params: { tuneName, patternName } }"
 			:title="readonly ? 'Show notes' : 'Edit notes'" v-b-tooltip.hover><fa icon="pen"/></router-link>
 		</li>
 		<li v-if="hasLocalChanges"><a href="javascript:" title="Revert modifications" v-b-tooltip.hover @click="restore()"><fa icon="eraser"/></a></li>
